requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pydantic-settings==2.10.1
jinja2==3.1.6
python-multipart==0.0.20
telethon==1.41.2
python-telegram-bot==22.3
pybit==5.12.0
openai==1.99.6
httpx==0.28.1
